[bd-q5efi,bd-bxngm,bd-t8y4c,bd-4rg59,bd-hi4fc] SoC continued + party repos + identity-auth migration
- AP 3.2.1: vendors + aging repo extraction (bd-q5efi)
- AR 6.2.2: domain module refinements (bd-t8y4c)
- Inventory 2.6.1: valuation, expiry, labels, cycle_count, status repo wiring (bd-bxngm)
- Party 3.2.0: 7 repo modules extracted from db layer (bd-4rg59)
- Identity-auth 1.9.1: migration 008 SoD scope/description columns (bd-hi4fc)

+++ modules/ar/src/domain/events.rs
@@ -9,8 +9,8 @@ pub async fn list_events<'e>(
     executor: impl PgExecutor<'e>,
     app_id: &str,
     query: &ListEventsQuery,
-    limit: i64,
-    offset: i64,
+    limit: i32,
+    offset: i32,
 ) -> Result<Vec<Event>, sqlx::Error> {
     let mut sql = String::from(
         "SELECT id, app_id, event_type, source, entity_type, entity_id, payload, created_at FROM ar_events WHERE app_id = $1",
